test: tests TaskRepository

+++ taskTracker/src/main/kotlin/repositories/TaskRepository.kt
@@ -5,19 +5,26 @@ import kotlinx.serialization.json.Json
 import models.Status
 import models.Task
 import java.io.File
+import java.io.FileNotFoundException
 import java.time.LocalDate
 
 class TaskRepository {
 
+    private val pathName = "src/main/resources"
     private var tasks: MutableList<Task>
 
     init {
         tasks = initializeList()
     }
 
-    fun save(task: Task) {
-        tasks.add(task)
-        writeJson()
+    fun save(task: Task): Boolean {
+        val existingTask = tasks.find { it.id == task.id }
+        if (existingTask == null) {
+            tasks.add(task)
+            writeJson()
+            return true
+        }
+        return false
     }
 
     fun findAll() = tasks
@@ -27,6 +34,8 @@ class TaskRepository {
     fun findAllInProgress() = tasks.filter { it.status == Status.PROGRESS }
 
     fun findAllDone() = tasks.filter { it.status == Status.DONE }
+
+    fun findById(id: Int) = tasks.find { it.id == id }
 
     fun updatedById(id: Int, newDescription: String): Boolean {
         val taskIndex = tasks.indexOfFirst { it.id == id }
@@ -61,11 +70,19 @@ class TaskRepository {
 
     private fun writeJson() {
         val json = Json.encodeToString(tasks)
-        File("src/main/resources/tasks.json").writeText(json)
+        File("${pathName}/tasks.json").writeText(json)
     }
 
     private fun initializeList(): MutableList<Task> {
-        val json = File("src/main/resources/tasks.json").readText()
-        return Json.decodeFromString(json)
+        try {
+            val json = File("${pathName}/tasks.json").readText()
+            return Json.decodeFromString(json)
+        } catch (e: FileNotFoundException) {
+            File("${pathName}/tasks.json").createNewFile()
+            val emptyListTask = mutableListOf<Task>()
+            val json = Json.encodeToString(emptyListTask)
+            File("${pathName}/tasks.json").writeText(json)
+            return Json.decodeFromString(json)
+        }
     }
 }
